♻️ 인터셉터, index.ts 파일 분리 (#81)
